Send audio data at the beginning of each 20ms to increase timing consistency
Previously the behavior was at frame start, render, send, then wait, which would potentially add a lot of variability to when the send actually occurred. The new behavior is send, render, wait, sending always occurring as close to the start of the 20ms window we can get and render/wait evenly filling out the rest of the 20ms.